Move content of temporary test file by overriding fixtures locally
